Update weather  home page

+++ learnflutter/lib/exercise/weather_app/models/weather_result.dart
@@ -134,7 +134,7 @@ class WeatherMainObject {
 }
 
 class WindObject {
-  double? speed;
+  int? speed;
   int? deg;
   double? gust;
